Change node address type to 16 bits

+++ src/lib/wkcomm/c/common/wkcomm.h
@@ -19,7 +19,7 @@
 
 // WuKong address. For now it's just a byte, but this will probably change.
 // When it does, we need to change the component-node map as well
-typedef uint8_t wkcomm_address_t;
+typedef uint16_t wkcomm_address_t;
 
 typedef struct wkcomm_received_msg {
 	wkcomm_address_t src;
